Modified Array Object component Addition/Deletion to handle all operators and objects
