+++ frontend/src/component/common/GuiConstants.js
@@ -3,3 +3,4 @@ import React from "react";
 export const USE_MOCK_DATA = true;
 export const FAKE_SUBMIT = true;
 export const BACKEND_URL = "http://localhost:8080/backend/rest";
+export const OPERATOR_NAME = "RHPAM";
